+++ scripts/constants.js
@@ -4,6 +4,7 @@ export const SETTINGS = {
   TARGET_HELPER: "targetHelper",
   TARGET_HELPER_AUTOMATIONS: "targetHelperAutomations",
   TARGET_HELPER_AUTOMATIONS_NPC_ONLY: "targetHelperAutomationsNpcOnly",
+  TARGET_HELPER_COLOR_SCHEME: "targetHelperColorScheme",
   REACH_CONTROL: "reachControl",
   REACH_DOORS: "reachControlDoors",
   REACH_DOOR_RANGE: "reachControlDoorRange",
@@ -13,6 +14,11 @@ export const SETTINGS = {
   REACH_STAIRWAYS_AFFECT_GM: "reachControlStairwaysAffectGM",
   REACH_TOKENS: "reachControlTokens",
   REACH_TOKEN_RANGE: "reachControlTokenRange"
+};
+
+export const TARGET_HELPER_COLOR_SCHEMES = {
+  DEFAULT: "default",
+  HIGH_CONTRAST: "highContrast"
 };
 
 export const REACH_RANGE = {